+++ src/MarlothStrategy.Simulation/Production/SignalValue.cs
@@ -7,35 +7,35 @@ namespace MarlothStrategy.Simulation.Production;
 /// </summary>
 public abstract record SignalValue(SignalTypeId TypeId)
 {
-    public sealed record Money(decimal Amount) : SignalValue(SignalTypes.Money)
+    public sealed record Money(int Amount) : SignalValue(SignalTypes.Money)
     {
-        public Money Add(decimal delta) => new(Amount + delta);
+        public Money Add(int delta) => new(Amount + delta);
 
-        public Money WithAmount(decimal amount) => new(amount);
+        public Money WithAmount(int amount) => new(amount);
     }
 
-    public sealed record Enchantments(decimal Amount) : SignalValue(SignalTypes.Enchantments)
+    public sealed record Enchantments(int Amount) : SignalValue(SignalTypes.Enchantments)
     {
-        public Enchantments Add(decimal delta) => new(Amount + delta);
+        public Enchantments Add(int delta) => new(Amount + delta);
 
-        public Enchantments WithAmount(decimal amount) => new(amount);
+        public Enchantments WithAmount(int amount) => new(amount);
     }
 
-    public decimal Quantity => this switch
+    public int Quantity => this switch
     {
         Money m => m.Amount,
         Enchantments e => e.Amount,
         _ => throw new InvalidOperationException($"Unknown signal value kind: {GetType().Name}."),
     };
 
-    public SignalValue WithQuantity(decimal amount) => this switch
+    public SignalValue WithQuantity(int amount) => this switch
     {
         Money => new Money(amount),
         Enchantments => new Enchantments(amount),
         _ => throw new InvalidOperationException($"Unknown signal value kind: {GetType().Name}."),
     };
 
-    public SignalValue AddQuantity(decimal delta) => WithQuantity(Quantity + delta);
+    public SignalValue AddQuantity(int delta) => WithQuantity(Quantity + delta);
 }
 
 public static class SignalTypes
